Usuniecie klasy Node i umieszczenie jej w klasie IntLinkedList, z powodu tego że jako jedyna jej używa
