+++ src/main/java/edu/kis/vh/nursery/DefaultCountingOutRhymer.java
@@ -7,22 +7,22 @@ public class DefaultCountingOutRhymer {
 	private static final int arraysize = 12;
 
 	private int[] numbers = new int[arraysize];
-	
+
 	public int total = initialtotal;
 
 	public int getTotal() {
 		return total;
 	}
-	
+
 	public void countIn(int in) {
 		if (!isFull())
 			numbers[++total] = in;
 	}
-	
+
 	public boolean callCheck() {
 		return total == initialtotal;
 	}
-	
+
 	public boolean isFull() {
 		return total == 11;
 	}
