+++ src/main/java/edu/kis/vh/nursery/HanoiRhymer.java
@@ -7,6 +7,7 @@ public class HanoiRhymer extends DefaultCountingOutRhymer {
 	public int reportRejected() {
 		return totalRejected;
 	}
+
 	@Override
 	public void countIn(int in) {
 		if (!callCheck() && in > prevValue())
